Update README.md
Trying to get "meta.TITLE", "meta.DESCRIPTION" etc. for a URL which is undefined in Redis, express would return a 500 since the "meta" variable itself was undefined.
This cleans up the view code a little bit and fixed the error - URLs still undefined in Redis will now get the default values.